requirements.txt
py-cord==2.4.1
flake8~=4.0
python-dotenv~=0.19
asyncio
